Turn ISlackAdapter interface into a class
