+++ tests/Adapters/Microsoft.Bot.Builder.Adapters.Slack.TestBot/SimpleSlackAdapterOptions.cs
@@ -2,7 +2,7 @@
 
 namespace Microsoft.Bot.Builder.Adapters.Slack.TestBot
 {
-    public class SimpleSlackAdapterOptions : ISlackAdapterOptions
+    public class SimpleSlackAdapterOptions : SlackAdapterOptions
     {
         public SimpleSlackAdapterOptions()
         {
@@ -13,30 +13,6 @@ namespace Microsoft.Bot.Builder.Adapters.Slack.TestBot
             this.VerificationToken = verificationToken;
             this.BotToken = botToken;
             this.ClientSigningSecret = signingSecret;
-        }
-
-        public string VerificationToken { get; set; }
-
-        public string ClientSigningSecret { get; set; }
-
-        public string BotToken { get; set; }
-
-        public string ClientId { get; set; }
-
-        public string ClientSecret { get; set; }
-
-        public string[] Scopes { get; set; }
-
-        public string RedirectUri { get; set; }
-
-        public Task<string> GetBotUserByTeam(string teamId)
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public Task<string> GetTokenForTeam(string teamId)
-        {
-            throw new System.NotImplementedException();
         }
     }
 }
